+++ src/core/server/graph/loaders/Auth.ts
@@ -1,8 +1,10 @@
 import DataLoader from "dataloader";
 
 import GraphContext from "coral-server/graph/context";
-import { GQLDiscoveredOIDCConfiguration } from "coral-server/graph/schema/__generated__/types";
+import { retrieveLastUsedAtTenantSSOKeys } from "coral-server/models/tenant";
 import { discoverOIDCConfiguration } from "coral-server/services/tenant";
+
+import { GQLDiscoveredOIDCConfiguration } from "coral-server/graph/schema/__generated__/types";
 
 export default (ctx: GraphContext) => ({
   discoverOIDCConfiguration: new DataLoader<
@@ -16,5 +18,8 @@ export default (ctx: GraphContext) => ({
       // long lived.
       cache: !ctx.disableCaching,
     }
+  ),
+  retrieveSSOKeyLastUsedAt: new DataLoader((kids: string[]) =>
+    retrieveLastUsedAtTenantSSOKeys(ctx.redis, ctx.tenant.id, kids)
   ),
 });
